backend/requirements.txt
flask
flask-cors
google-cloud-spanner
vertexai
google-cloud-aiplatform
gunicorn

